requirements.txt
# Core dependencies
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pyttsx3>=2.90
openai>=0.27.8
yt-dlp>=2023.7.6
replicate>=0.8.3
Pillow>=10.0.0
requests>=2.31.0
torch>=2.0.0
transformers>=4.30.0

# Development dependencies
black>=23.7.0
flake8>=6.1.0
pytest>=7.4.0